attempt to create e2e tests
