+++ package.json
@@ -4,7 +4,8 @@
   "description": "A micro library to get, set, and delete Consul K/V",
   "main": "consul.js",
   "scripts": {
-    "test": "mocha && eslint *.js test/*.js"
+    "test": "mocha --recursive test/unit --exit && eslint *.js test/*.js",
+    "test:e2e": "mocha --recursive test/e2e"
   },
   "author": "Mike Ball",
   "license": "MIT",
