+++ test/consul_e2e_test.js
@@ -1,50 +1,24 @@
 const assert = require('assert');
-const nock = require('nock');
-const Consul = require('../consul');
-const consulHost = 'my-consul.com';
+const Consul = require('../../consul');
 const consul = new Consul({
-  host: consulHost,
-  token: 'my-token',
-  tlsCert: 'cert',
-  tlsKey: 'key'
+  host: 'localhost',
+  protocol: 'http'
 });
 
-const host = `https://${consulHost}:8500`;
-const endpoint = '/v1/kv/my/key?token=my-token';
+describe('set', () => {
+  it('sets the key/value it is passed', (done) => {
+    consul.set('my/key', 'my-value')
+      .then(result => {
+        assert.equal(result, true);
 
-function mockGet(query) {
-  const url = query ? `${endpoint}${query}` : endpoint;
-
-  return nock(host)
-    .get(url)
-    .reply(200, [{
-      Value: 'bXktdmFsdWU='
-    }]);
-}
-
-function mockGet404() {
-  return nock(host)
-    .get(endpoint)
-    .reply(404, []);
-}
-
-function mockPut() {
-  return nock(host)
-    .put(endpoint, 'my-value')
-    .reply(200, 'true');
-}
-
-function mockDelete() {
-  return nock(host)
-    .delete(endpoint)
-    .reply(200, 'true');
-}
+        done();
+      });
+  });
+});
 
 describe('get', () => {
   describe('when it 200s', () => {
     it('returns the decoded value for the key it is passed', (done) => {
-      mockGet();
-
       consul.get('my/key')
         .then(val => {
           assert.equal(val.value, 'my-value');
@@ -54,8 +28,6 @@ describe('get', () => {
     });
 
     it('returns the status code of the response', (done) => {
-      mockGet();
-
       consul.get('my/key')
         .then(val => {
           assert.equal(val.responseStatus, 200);
@@ -66,8 +38,6 @@ describe('get', () => {
 
     describe('when it is passed the option to recursively return the entire subtree', () => {
       it('adds "?recurse" to the request it makes', (done) => {
-        mockGet('&recurse');
-
         consul.get('my/key', { recurse: true })
           .then(val => {
             assert.equal(val.value, 'my-value');
@@ -77,22 +47,7 @@ describe('get', () => {
       });
     });
 
-    describe('when it is passed the option to speicfy the datacenter to query', () => {
-      it('adds "?dc=my-dc" to the request it makes', (done) => {
-        mockGet('&dc=my-dc');
-
-        consul.get('my/key', { dc: 'my-dc' })
-          .then(val => {
-            assert.equal(val.value, 'my-value');
-
-            done();
-          });
-      });
-    });
-
     it('returns the body of the response', (done) => {
-      mockGet();
-
       consul.get('my/key')
         .then(val => {
           assert.equal(val.responseBody[0].Value, 'bXktdmFsdWU=');
@@ -104,9 +59,7 @@ describe('get', () => {
 
   describe('when it 404s', () => {
     it('throws an error', (done) => {
-      mockGet404();
-
-      consul.get('my/key')
+      consul.get('my/foo')
         .catch(err => {
           assert.equal(err.code, 'ERR_BAD_REQUEST');
 
@@ -116,23 +69,8 @@ describe('get', () => {
   });
 });
 
-describe('set', () => {
-  it('sets the key/value it is passed', (done) => {
-    mockPut();
-
-    consul.set('my/key', 'my-value')
-      .then(result => {
-        assert.equal(result, true);
-
-        done();
-      });
-  });
-});
-
 describe('delete', () => {
   it('deletes the key/value it is passed', (done) => {
-    mockDelete();
-
     consul.delete('my/key')
       .then(result => {
         assert.equal(result, true);
